Fix ir.ts handling of missing attributes
We can't use a ? unless we fix things on the Python side because `undefined`
turns into an entirely missing entry. If we want to turn it into `None` we need
to either put a `null` in javascript or set `None` as the default value in
`ir.py`. I removed a lot of the `?`'s because we don't use them, but
qualifiedName and sourcefilename do sometimes seem to go missing right now.
Maybe we can fix that later.