Move most of the logic from typedoc.py into typescript  (#119)
Prior to this PR we have typedoc generate a json serialization of its IR, then
we load this JSON in Python using a Pydantic model.

There are a few problems with this: The serialization format for typedoc is
unstable across minor versions and the schema is complicated and mutually
recursive. When the updates happen, Pydantic generates very complicated schema
violation errors and it's difficult to understand what specifically went wrong
or how to fix it. The schema is also far from the location in the code that uses
it so it can be hard to understand what the intent of the information needed is.
It's also hard to develop new features because it's not very easy to explore the
serialized json to find information, whereas in typescript one can easily
navigate to the definitions of typedoc objects. Lastly, the typedoc IR has not
just data but also methods which often implement pretty useful logic but we
can't use any of them.

This reimplements most of the logic in typedoc.py to typescript. The new
transfer format between typescript and Python is ir.py which gained logic to
serialize and deserialize it. I used cattrs for this purpose instead of pydantic
because cattrs allows us to add serialization to an existing attrs class without
changing its API. If we changed the classes in ir.py into Pydantic classes all
of the spots where we instantiate them in Python would have to change. It took
about 70 lines of added code to do this. Using Pydantic would involve less added
code but more changes to existing code. A big advantage of ir.py is that it is
not a recursive data format so when schema errors happen they are much easier to
reason about.