Fix handling XRefs for Reference types that refer to a SymbolId (#132)
I don't exactly understand how ReferenceTypes work in typedoc but they seem to
have two options:

1. The referent is a documented object and so there is a `reflection` field that
   points to the documentation reflection for the target.
2. It is not a documented object so the target is a SymbolId. This contains some
   information about the original typescript symbol, but not all of it.

So if we see a Reference it should definitely point to an external XRef. But
sometimes it seems to be that the SymbolId is also an internal object,
particularly when it's something being reexported.

This reference rendering code is still not quite right but at least now it
passes the added test.