Add some typescript tests
So far these more or less duplicate a subset of the tests in
`test_typedoc_analysis` so they don't provide that much added value because
they are still basically integration tests with respect to the typescript
logic. But this is a first step and will allow finer grained typescript tests
in the future.